+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-05-17
+
 ### Added
 
 - Add getters for all fields in `Opening` [#25]
@@ -43,5 +45,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#13]: https://github.com/dusk-network/merkle/issues/13
 
 <!-- VERSIONS -->
-[Unreleased]: https://github.com/dusk-network/merkle/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/dusk-network/merkle/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/dusk-network/merkle/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/dusk-network/merkle/releases/tag/v0.1.0